updated to bind scope.toggled instead of attr.expanded
The original author did most of the ground work, but for my needs I wanted to bind the toggled state to a parent scope. I modified the directive names to reflect their roles in the way I saw most accurate. Added a timeout function to slide the content in or out based on the value passed to the directive through scope.toggled.  I think this is a more "angular" way, since now I can bind this value in two directions; this allows me to save the clients toggled state. I also arranged the sliding code into an encapsulated function to improve readability and code reuse.